# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwiftLingo - Telegram Mini App translation marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
...
  /bids/{bidId}:
    patch:
      operationId: updateBid
      tags: [bids]
      summary: Update a bid (translator only, own bid)
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidUpdate"
      responses:
        "200":
          description: Updated bid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
    delete:
      operationId: withdrawBid
      tags: [bids]
      summary: Withdraw a bid (translator only)
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Withdrawn
  /bids/{bidId}/accept:
    post:
      operationId: acceptBid
      tags: [bids]
      summary: Client accepts a bid, creating a contract
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Contract created, pending payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts for current user (as client or translator)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractSummary"
  /contracts/{contractId}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract detail
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          description: Not found
...
  /contracts/{contractId}/approve:
    post:
      operationId: approveDelivery
      tags: [contracts]
      summary: Client approves delivery, releasing payment
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
  /contracts/{contractId}/dispute:
    post:
      operationId: disputeContract
      tags: [contracts]
      summary: Raise a dispute on a contract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "200":
          description: Contract in dispute
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, telegramId, role, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, translator, admin]
        preferredLanguage:
          type: string
          enum: [en, uz, ru]
        isTranslatorApproved:
          type: boolean
        createdAt:
          type: string
...
    TranslatorProfile:
      type: object
      required: [id, userId, pricePerWord, sourceLanguages, targetLanguages, specializations, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bio:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["integer", "null"]
        pricePerWord:
          type: number
        sourceLanguages:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        completedJobsCount:
          type: integer
        createdAt:
          type: string
...
    Job:
      type: object
      required: [id, clientId, title, sourceLang, targetLang, status, budgetMin, budgetMax, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        sourceLang:
          type: string
        targetLang:
          type: string
        wordCount:
          type: ["integer", "null"]
        budgetMin:
          type: number
        budgetMax:
          type: number
        deliveryType:
          type: string
          enum: [standard, express, urgent]
        specialization:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        bidsCount:
          type: integer
        createdAt:
          type: string
...
    Bid:
      type: object
      required: [id, jobId, translatorId, amount, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        translatorId:
          type: integer
        amount:
          type: number
        coverLetter:
          type: ["string", "null"]
        deliveryDays:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        translator:
          $ref: "#/components/schemas/TranslatorProfile"
        translatorUser:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
...
    BidUpdate:
      type: object
      properties:
        amount:
          type: number
        coverLetter:
          type: string
        deliveryDays:
          type: integer
    Contract:
      type: object
      required: [id, jobId, clientId, translatorId, agreedPrice, status, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        clientId:
          type: integer
        translatorId:
          type: integer
        bidId:
          type: integer
        agreedPrice:
          type: number
        platformFee:
          type: number
        translatorPayout:
          type: number
        status:
          type: string
          enum: [pending_payment, active, delivered, completed, cancelled, disputed]
        paymentStatus:
          type: string
          enum: [pending, confirmed, released, refunded]
        deliveryNote:
          type: ["string", "null"]
        disputeReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ContractSummary:
      allOf:
        - $ref: "#/components/schemas/Contract"
        - type: object
          properties:
            job:
              $ref: "#/components/schemas/Job"
            client:
              $ref: "#/components/schemas/User"
            translator:
              $ref: "#/components/schemas/User"
    ContractDetail:
      allOf:
        - $ref: "#/components/schemas/ContractSummary"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
            payment:
              $ref: "#/components/schemas/Payment"
            review:
              $ref: "#/components/schemas/Review"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/ContractStatusHistory"
...
    DisputeInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    Message:
      type: object
      required: [id, contractId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        contractId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        fileType:
          type: ["string", "null"]
        sender:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, contractId, amount, status, createdAt]
      properties:
        id:
          type: integer
        contractId:
          type: integer
        amount:
          type: number
        platformFee:
          type: number
        translatorPayout:
          type: number
        status:
          type: string
          enum: [pending, confirmed, released, refunded]
        paymentMethod:
          type: ["string", "null"]
        createdAt:
          type: string
        confirmedAt:
          type: ["string", "null"]
        releasedAt:
          type: ["string", "null"]
...
    Review:
      type: object
      required: [id, contractId, clientId, translatorId, rating, createdAt]
      properties:
        id:
          type: integer
        contractId:
          type: integer
        clientId:
          type: integer
        translatorId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: ["string", "null"]
        client:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
...
    ContractStatusHistory:
      type: object
      required: [id, contractId, status, createdAt]
      properties:
        id:
          type: integer
        contractId:
          type: integer
        status:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
...